+++ android/src/main/java/com/capacitorjs/community/plugins/bluetoothle/BluetoothLe.kt
@@ -441,6 +441,30 @@ class BluetoothLe : Plugin() {
     }
 
     @PluginMethod
+    fun getBondedDevices(call: PluginCall) {
+        assertBluetoothAdapter(call) ?: return
+
+        val bluetoothManager = activity.getSystemService(Context.BLUETOOTH_SERVICE) as BluetoothManager
+        val bluetoothAdapter = bluetoothManager.adapter
+
+        if (bluetoothAdapter == null) {
+            call.reject("Bluetooth is not supported on this device")
+            return
+        }
+
+        val bondedDevices = bluetoothAdapter.bondedDevices
+        val bleDevices = JSArray()
+
+        bondedDevices.forEach { device ->
+            bleDevices.put(getBleDevice(device))
+        }
+
+        val result = JSObject()
+        result.put("devices", bleDevices)
+        call.resolve(result)
+    }
+
+    @PluginMethod
     fun connect(call: PluginCall) {
         outputStream = null
         bluetoothSocket = null